+++ CHANGELOG.md
@@ -14,9 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - valgrind
 - Gitlab CI
 - pkg-config and version scripts for device token library
+- more fields in configuration file - it can hold the whole config now
 
 ### Changed
 - split server CLI option to server and port options
+- CLI options - names, arguments and their descriptions updated
 
 ## [1.4] - 2021-05-31
 ### Added